Move ssg/stig.py back to utils/xccdf2csv-stig.py
